pointing_device_gestures: Q8 fixed point cursor glide
Floating point is actually faster on AVR; cursor_glide_start() is around
33% faster with the call to hypotf() (~218us vs. ~327us), cursor_glide()
is about equal between floating point and fixed point.
Fixed point is much faster on a CPU with native 32-bit words and no FPU,
such as ARM Cortex-M0. (~55us vs. ~18us for cursor_glide_start on an M0)
Did not have a microcontroller board with Cortex-M4F to test performance
with FPU.

+++ quantum/pointing_device_gestures.h
@@ -27,7 +27,7 @@ typedef struct {
 typedef struct {
     uint16_t          trigger_px;
     uint16_t          coef;
-    float             v0;
+    int32_t           v0;
     int32_t           x;
     int32_t           y;
     uint16_t          z;
